Isolated Posts to be non-reactive on post_edit.
The idea here is to isolate the call to findOne() so that it stops
being reactive once data is found. The reason being that server scoring
updates will 'change' the post, and meteor will overwrite the user's
input, even if preserve is set. 
Kind of like {{#constant}}, except it waits for data.